refator: Change file structure, Add test code
