+++ Dockerfile.dev
@@ -1,18 +1,23 @@
-# Python 3.9 slim 이미지 사용
+# Python 3.12 slim 이미지 사용
 FROM python:3.12-slim
 
 # 작업 디렉토리 설정
 WORKDIR /app
 
-# 필요한 패키지를 설치
+# 필요한 패키지 설치
 COPY requirements.txt .
 
-RUN pip install --no-cache-dir  -r requirements.txt
+RUN pip install --no-cache-dir -r requirements.txt
 
-# 필요한 스크립트 및 파일 복사
-COPY test_review_code.py .
+# 소스 코드 및 테스트 코드 복사
+COPY src/ ./src
+COPY tests/ ./tests
 
-COPY main.py .
+# 환경 변수 설정 (필요한 경우)
+ENV PATH="/usr/local/bin:${PATH}"
 
-# 스크립트 실행에 필요한 기본 명령어 설정
-CMD ["python", "test_review_code.py"]
+# 환경 변수 설정 (PYTHONPATH 설정)
+ENV PYTHONPATH="${PYTHONPATH}:/app/src"
+
+# 테스트 실행을 위한 기본 명령어 설정
+CMD ["python3", "-m", "unittest", "discover", "-s", "tests", "-p", "test_*.py"]
